GRAF-87 Auth Guards & other misc changes (#169)
* GRAF-87 renamed time log folder

* GRAF-87 renamed meeting notes folder

* GRAF-87 added 1st semester project report

* GRAF-87 one time temp change to ci

* GRFA-87 script change

* GRAF-87 add draft prod.env file

* GRAF-87 add filebeat

* ran lint

* GRAF-87 prettierignore upadte

* GRAF-87 naming change

* GRAF-87 spelling mistake fix

* GRAF-87 added auth guards to category

* GRAF-87 Added AuthGuards to comment module

* GRAF-87 added auth guards to artist module

* GRAF-87 change yaml to yml

* GRAF-87 undoing temp ci changes

* GRAF-87 working like functionality

+++ src/artist/artist.service.ts
@@ -1,14 +1,23 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { Request } from 'express';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateArtistDto } from './dto/request/create-artist.dto';
 import { GraffitiEntry } from './dto/request/graffiti-entry.dto';
 import { UpdateArtistDto } from './dto/request/update-artist.dto';
-
+import { AuthService } from '../auth/auth.service';
 @Injectable()
 export class ArtistService {
-	constructor(private prisma: PrismaService) {}
+	constructor(
+		private prisma: PrismaService,
+		private authService: AuthService,
+	) {}
 
-	async create(createArtistDto: CreateArtistDto) {
+	async create(createArtistDto: CreateArtistDto, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
 		return await this.prisma.artist.create({
 			data: {
 				name: createArtistDto.name,
@@ -35,14 +44,48 @@ export class ArtistService {
 		return entity;
 	}
 
-	async addGraffitiToArtist(id: number, request: GraffitiEntry) {
+	async addGraffitiToArtist(
+		id: number,
+		entry: GraffitiEntry,
+		request: Request,
+	) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
 		let entity = await this.prisma.artist.update({
 			where: {
 				id: id,
 			},
 			data: {
 				graffitis: {
-					create: request.graffitiIds.map((graffitiId) => ({
+					create: entry.graffitiIds.map((graffitiId) => ({
+						graffitiId: graffitiId,
+					})),
+				},
+			},
+		});
+		return entity;
+	}
+
+	async removeGraffitiFromArtist(
+		id: number,
+		entry: GraffitiEntry,
+		request: Request,
+	) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
+		let entity = await this.prisma.artist.update({
+			where: {
+				id: id,
+			},
+			data: {
+				graffitis: {
+					deleteMany: entry.graffitiIds.map((graffitiId) => ({
 						graffitiId: graffitiId,
 					})),
 				},
@@ -52,33 +95,26 @@ export class ArtistService {
 		return entity;
 	}
 
-	async removeGraffitiFromArtist(id: number, request: GraffitiEntry) {
-		let entity = await this.prisma.artist.update({
-			where: {
-				id: id,
-			},
-			data: {
-				graffitis: {
-					deleteMany: request.graffitiIds.map((graffitiId) => ({
-						graffitiId: graffitiId,
-					})),
-				},
-			},
-		});
-		console.log(entity);
-		return entity;
-	}
+	async update(id: number, updateArtistDto: UpdateArtistDto, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
 
-	async update(id: number, request: UpdateArtistDto) {
 		return await this.prisma.artist.update({
 			where: {
 				id: id,
 			},
-			data: request,
+			data: updateArtistDto,
 		});
 	}
 
-	async delete(id: number) {
+	async delete(id: number, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
 		return await this.prisma.artist.delete({
 			where: {
 				id: id,
